refactor: content 도메인을 기능별 패키지 구조로 분리
- classroom과 동일한 패턴으로 content 루트 파일들을 패키지화.
- controller, dependency, repository, schema, service 폴더로 분리해
- IDE에서 폴더/파일 아이콘이 구분되어 구조 파악이 빠르다.
