+++ app/domains/content/service/service.py
@@ -1,5 +1,5 @@
 from app.domains.content.models import Lesson, Unit
-from app.domains.content.repository import ContentCatalogRepository
+from app.domains.content.repository.repository import ContentCatalogRepository
 
 
 class ContentCatalogService:
