+++ app/domains/progress/router.py
@@ -3,8 +3,8 @@ from fastapi import APIRouter, Depends
 from app.domains.auth.dependencies import get_current_student
 from app.domains.auth.dependencies import get_current_user
 from app.domains.auth.models import User
-from app.domains.content.dependencies import get_content_catalog_service
-from app.domains.content.service import ContentCatalogService
+from app.domains.content.dependency.dependencies import get_content_catalog_service
+from app.domains.content.service.service import ContentCatalogService
 from app.domains.progress.dependencies import get_learning_record_service
 from app.domains.progress.schemas import (
     LearningRecordResponse,
