+++ tests/learning/test_content_catalog_controller.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 
 from app.domains.auth.models import User
 from app.domains.content.models import Lesson, Unit
-from app.domains.content.router import get_content_lesson, get_content_units
+from app.domains.content.controller.router import get_content_lesson, get_content_units
 
 
 def _user() -> User:
