+++ tests/structure/test_domain_structure.py
@@ -31,7 +31,7 @@ def test_domain_packages_are_importable():
         "app.domains.instruction.repository",
         "app.domains.instruction.schemas",
         "app.domains.instruction.service",
-        "app.domains.content.router",
+        "app.domains.content.controller.router",
         "app.domains.content.stage1.router",
         "app.domains.content.stage1.schemas",
         "app.domains.content.stage1.service",
@@ -42,9 +42,9 @@ def test_domain_packages_are_importable():
         "app.domains.content.stage3.router",
         "app.domains.content",
         "app.domains.content.models",
-        "app.domains.content.repository",
-        "app.domains.content.schemas",
-        "app.domains.content.service",
+        "app.domains.content.repository.repository",
+        "app.domains.content.schema.schemas",
+        "app.domains.content.service.service",
         "app.domains.progress.models",
         "app.domains.progress.repository",
         "app.domains.progress.schemas",
